#2 Removing abstract class for version resolver of Chrome and Chromium
