+++ src/Browser/GoogleChrome/VersionResolver.php
@@ -6,15 +6,62 @@ namespace DBrekelmans\BrowserDriverInstaller\Browser\GoogleChrome;
 
 use DBrekelmans\BrowserDriverInstaller\Browser\BrowserName;
 use DBrekelmans\BrowserDriverInstaller\Browser\VersionResolver as VersionResolverInterface;
-use DBrekelmans\BrowserDriverInstaller\Browser\VersionResolverChrome;
+use DBrekelmans\BrowserDriverInstaller\CommandLine\CommandLineEnvironment;
+use DBrekelmans\BrowserDriverInstaller\Exception\NotImplemented;
+use DBrekelmans\BrowserDriverInstaller\OperatingSystem\OperatingSystem;
+use DBrekelmans\BrowserDriverInstaller\Version;
+use RuntimeException;
 
 use function Safe\sprintf;
 
-final class VersionResolver extends VersionResolverChrome implements VersionResolverInterface
+final class VersionResolver implements VersionResolverInterface
 {
-    protected function getMacOSCommandLineForVersion(string $path): string
+    private CommandLineEnvironment $commandLineEnvironment;
+
+    public function __construct(CommandLineEnvironment $commandLineEnvironment)
     {
-        return sprintf('%s/Contents/MacOS/Google\ Chrome --version', $path);
+        $this->commandLineEnvironment = $commandLineEnvironment;
+    }
+
+    public function from(OperatingSystem $operatingSystem, string $path): Version
+    {
+        if ($operatingSystem->equals(OperatingSystem::LINUX())) {
+            return $this->getVersionFromCommandLine(sprintf('%s --version', $path));
+        }
+
+        if ($operatingSystem->equals(OperatingSystem::MACOS())) {
+            return $this->getVersionFromCommandLine(
+                sprintf('%s/Contents/MacOS/Google\ Chrome --version', $path)
+            );
+        }
+
+        if ($operatingSystem->equals(OperatingSystem::WINDOWS())) {
+            return $this->getVersionFromCommandLine(
+                sprintf('wmic datafile where name="%s" get Version /value', $path)
+            );
+        }
+
+        throw NotImplemented::feature(
+            sprintf(
+                'Resolving version on %s',
+                $operatingSystem->getValue()
+            )
+        );
+    }
+
+    private function getVersionFromCommandLine(string $command): Version
+    {
+        try {
+            $commandOutput = $this->commandLineEnvironment->getCommandLineSuccessfulOutput($command);
+
+            return Version::fromString($commandOutput);
+        } catch (RuntimeException $exception) {
+            throw new RuntimeException(
+                'Version could not be determined.',
+                0,
+                $exception
+            );
+        }
     }
 
     public function supports(BrowserName $browserName): bool
